requirements.txt
pyTelegramBotAPI
python-dotenv
wakeonlan